+++ rumqttc-v5/examples/async_auth_scram.rs
@@ -151,7 +151,7 @@ async fn main() -> Result<(), Box<dyn Error>> {
     mqttoptions.set_authentication_method(Some("SCRAM-SHA-256".to_string()));
     mqttoptions.set_authentication_data(client_first);
     mqttoptions.set_auth_manager(authmanager.clone());
-    let (client, mut eventloop) = AsyncClient::new(mqttoptions, 10);
+    let (client, mut eventloop) = AsyncClient::builder(mqttoptions).capacity(10).build_async();
 
     let (tx, rx) = bounded(1);